move sql rewrite engine out of rewriter package
